modify node config json flag and move Config from node to util
