+++ cmd/node/cmd/config.go
@@ -11,42 +11,18 @@ import (
 	"math/big"
 	"path/filepath"
 
+	"github.com/seeleteam/go-seele/cmd/util"
 	"github.com/seeleteam/go-seele/common"
 	"github.com/seeleteam/go-seele/core"
 	"github.com/seeleteam/go-seele/crypto"
 	"github.com/seeleteam/go-seele/log/comm"
-	"github.com/seeleteam/go-seele/metrics"
 	"github.com/seeleteam/go-seele/node"
 	"github.com/seeleteam/go-seele/p2p"
 )
 
-// Config is the Configuration of node
-type Config struct {
-	//Config is the Configuration of log
-	LogConfig comm.LogConfig `json:"log"`
-
-	// basic config for Node
-	BasicConfig node.BasicConfig `json:"basic"`
-
-	// The configuration of p2p network
-	P2PConfig p2p.Config `json:"p2p"`
-
-	// HttpServer config for http server
-	HTTPServer node.HTTPServer `json:"httpServer"`
-
-	// The configuration of websocket rpc service
-	WSServerConfig node.WSServerConfig `json:"wsserver"`
-
-	// metrics config info
-	MetricsConfig *metrics.Config `json:"metrics"`
-
-	// genesis config info
-	GenesisConfig core.GenesisInfo `json:"genesis"`
-}
-
 // GetConfigFromFile unmarshals the config from the given file
-func GetConfigFromFile(filepath string) (*Config, error) {
-	var config Config
+func GetConfigFromFile(filepath string) (*util.Config, error) {
+	var config util.Config
 	buff, err := ioutil.ReadFile(filepath)
 	if err != nil {
 		return &config, err
@@ -85,7 +61,7 @@ func LoadConfigFromFile(configFile string, accounts string) (*node.Config, error
 }
 
 // CopyConfig copy Config from the given config
-func CopyConfig(cmdConfig *Config) *node.Config {
+func CopyConfig(cmdConfig *util.Config) *node.Config {
 	config := &node.Config{
 		BasicConfig:    cmdConfig.BasicConfig,
 		LogConfig:      cmdConfig.LogConfig,
@@ -99,7 +75,7 @@ func CopyConfig(cmdConfig *Config) *node.Config {
 }
 
 // GetP2pConfig get P2PConfig from the given config
-func GetP2pConfig(cmdConfig *Config) (p2p.Config, error) {
+func GetP2pConfig(cmdConfig *util.Config) (p2p.Config, error) {
 	if cmdConfig.P2PConfig.PrivateKey == nil {
 		key, err := crypto.LoadECDSAFromString(cmdConfig.P2PConfig.SubPrivateKey) // GetP2pConfigPrivateKey get privateKey from the given config
 		if err != nil {
